Feat: Coursel divided into different file
